fix: Fix python setup
Do not symlink python3 to the uv python version, as it will break the
system python setup.

Also move the setup of Python and uv upwards in the Dockerfile to speed
up the build process, since we may more often change the Python setup
then the other parts of the Dockerfile.