build: Add build arg for Ollama installation
Add `WITH_OLLAMA` build argument to the base image to make Ollama
installation optional during the build process. This allows for
smaller images when Ollama is not needed.

The installation is now wrapped in a conditional check that only
installs Ollama when `WITH_OLLAMA=true` is passed as a build
argument. By default, Ollama will not be installed.